[FEATURE] Interface of trace
We implement data type and functions of tracer.

Some key features we introced to tracer:
(1) There are two types of mode during trace. The difference of the two
is that where the data will be record, file or struct in memory.
(2) Since we have introduced a global variable to control the process,
only one tracing process will be allowed under one system at a time.

Signed-off-by: YushuoEdge <y-shuo@sjtu.edu.cn>