[FEATURE] Implement tracer class
We implemented the complete tracer class, including constructs,
destructors, all member functions and so on.

Special attention should be paid:
(1) A global pointer in trace.h points to the only existing tracer.
(2) Since the file and objects are one-to-one correspondence, we implemente
a fstream as global variable in activity_trace_async.cpp. Also, we use a
traceCount, representing how many threads are tracing now, to control opening
and closing the file.

Signed-off-by: YushuoEdge <y-shuo@sjtu.edu.cn>