[FEATURE] writeFileActivity and storeValueActivity
Writing record to file and to the traceData strcut are implemented
through functions writeFileActivity and storeValueActivity.

In writeFileActivity, we add a new feature, translating cbid to callback function
name. This is for driver and runtime api kinds activity.

As for storing values process, we recategorized traceData into three kinds of classes:
(1) Runtime -- traceData_st, including memset, memcpy and kernel.
(2) API -- traceData_api, including driver_api, runtime_api.
(3) Overhead -- traceData_oh, including overhead.
And we add a dataTypeFlag member variable for tracer. This is used to select which kinds of
above classes to store, as this process will cost a lot of memory.

Signed-off-by: YushuoEdge <y-shuo@sjtu.edu.cn>