[FEATURE] Build Amanda Profiler and Metric Info
We built the framework of amanda profiler in this commit, including
the user interface, metrics layer and lower support. And we implement
it in pytorch in this commit.

KernelInfo is an example metric. It can print opIndex, opName, kernelIndex,
kernelType, kernelName, cudaLauchKernelApiTime, kernelExecutionTime, dramRead,
dramWrite, opCount, cyclesElapsed by integrate the data collected by tracer
and counter.

Signed-off-by: YushuoEdge <y-shuo@sjtu.edu.cn>