[BUGFIX] Mapping mechanism
In this commit, two kinds of mapping problems are fixed:
(1) OP-KERNEL Mapping. The mapping mechanisms has been rewritten
here. The previous one would cause some errors.
(2) Tracer-Counter in tensorflow. We found that when tracer and
counter run on a computational graph in a process, they interact with
each other. And the kernel is not exactly same when running some ops.
So we define metrics xxInfoTracer and xxInfoCounter to measure
separately.

Signed-off-by: YushuoEdge <y-shuo@sjtu.edu.cn>